+++ spec/src/mock/circular_parent.ts
@@ -1,0 +1,7 @@
+import { Serialize } from '../../../';
+
+import Child from './circular_child';
+
+export default class Parent {
+    @Serialize() public child: Child;
+}