+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed a bug with circular module dependencies
+
 v1.2.1
 ------